+++ hive_generator/lib/src/adapter_builder/adapter_builder.dart
@@ -49,19 +49,15 @@ abstract class AdapterBuilder {
   /// TODO: Document this!
   final List<AdapterField> setters;
 
-  /// Converters from [AdapterSpec.converters]
-  final List<DartObject> specConverters;
-
   /// Converters from [GenerateAdapters.converters]
-  final List<DartObject> globalConverters;
+  final List<DartObject> converters;
 
   /// TODO: Document this!
   const AdapterBuilder(
     this.cls,
     this.getters, {
     this.setters = const <AdapterField>[],
-    this.specConverters = const [],
-    this.globalConverters = const [],
+    this.converters = const [],
   });
 
   /// TODO: Document this!