+++ hive_generator/lib/src/generator/adapters_generator.dart
@@ -96,8 +96,7 @@ class AdaptersGenerator extends GeneratorForAnnotation<GenerateAdapters> {
         typeId: schemaType.typeId,
         schema: schemaType,
         ignoredFields: spec.ignoredFields,
-        specConverters: spec.converters,
-        globalConverters: revived.converters,
+        converters: revived.converters,
       );
 
       content.write(result.content);