+++ hive/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.20.0
 
 - Adds `HiveConverter` for custom field conversion during TypeAdapter generation (similar to `json_serializable`'s `JsonConverter`)
-- Adds `converters` parameter to `GenerateAdapters` and `AdapterSpec`
+- Adds `converters` parameter to `GenerateAdapters`
 
 ## 2.19.3